TradeApp (#1598)
* Add global styles

* Add global styles

* Rename TradeLayout to generic layout

* Create its own structure per app
