+++ src/app/TradeApp/pages/Trading.stories.tsx
@@ -1,10 +1,10 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta } from '@storybook/react/types-6-0'
 
-import { Trading } from 'pages/Trading'
+import { Trading } from './Trading'
 
 export default {
-  title: 'Pages/Trading',
+  title: 'TradeApp/Trading',
   component: Trading,
 } as Meta
