+++ src/app/TradeApp/pages/Trading.tsx
